requirements.txt
# ═══════════════════════════════════════════════════════════════
# TRAVEL WITH ALINA - Production Requirements
# Для 500+ пользователей
# ═══════════════════════════════════════════════════════════════

# Веб-фреймворк
Flask==3.0.0
Werkzeug==3.0.1

# Production серверы
waitress==2.1.2              # Для Windows (основной)
gunicorn==21.2.0             # Для Linux (опционально)

# OpenAI API
openai==1.12.0

# HTTP клиенты
requests==2.31.0
aiohttp==3.9.3
urllib3==2.1.0

# Утилиты
python-dotenv==1.0.0         # Для .env файлов
pytz==2024.1                 # Часовые пояса

# Парсинг и обработка данных
beautifulsoup4==4.12.3       # Для парсинга HTML (если нужно)
lxml==5.1.0                  # XML парсер

# Дополнительно (опционально)
# redis==5.0.1               # Для кэша в Redis (если нужно масштабирование)
# python-telegram-bot==20.7  # Если нужен отдельный Telegram бот

